requirements.txt
flask
flask_WTF
Flask-login
Flask_SQLalchemy
gunicorn
